refactor(host): http_server to explicit address file
