+++ crates/host/src/wasmbus/providers/http_server/address.rs
@@ -1,37 +1,31 @@
 use core::net::SocketAddr;
-use core::str::FromStr as _;
 
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use anyhow::{bail, Context as _};
+use anyhow::Context as _;
 use http::header::HOST;
 use http::uri::Scheme;
 use http::Uri;
 use http_body_util::BodyExt as _;
 use hyper_util::rt::{TokioExecutor, TokioIo};
-use nkeys::XKey;
-use tokio::sync::{broadcast, Mutex, RwLock};
+use tokio::sync::{Mutex, RwLock};
 use tokio::task::JoinSet;
 use tokio::time::Instant;
 use tracing::{error, info_span, instrument, trace_span, Instrument as _, Span};
-use wasmcloud_core::InterfaceLinkDefinition;
-use wasmcloud_provider_http_server::{default_listen_address, load_settings, ServiceSettings};
-use wasmcloud_provider_sdk::provider::{
-    handle_provider_commands, receive_link_for_provider, ProviderCommandReceivers,
-};
-use wasmcloud_provider_sdk::{LinkConfig, LinkDeleteInfo, ProviderConnection};
+use wasmcloud_provider_http_server::{load_settings, ServiceSettings};
+use wasmcloud_provider_sdk::{LinkConfig, LinkDeleteInfo};
 use wasmcloud_tracing::KeyValue;
 use wrpc_interface_http::ServeIncomingHandlerWasmtime as _;
 
 use crate::wasmbus::{Component, InvocationContext};
 
-struct Provider {
-    address: SocketAddr,
-    components: Arc<RwLock<HashMap<String, Arc<Component>>>>,
-    links: Mutex<HashMap<Arc<str>, HashMap<Box<str>, JoinSet<()>>>>,
-    lattice_id: Arc<str>,
-    host_id: Arc<str>,
+pub(crate) struct Provider {
+    pub(crate) address: SocketAddr,
+    pub(crate) components: Arc<RwLock<HashMap<String, Arc<Component>>>>,
+    pub(crate) links: Mutex<HashMap<Arc<str>, HashMap<Box<str>, JoinSet<()>>>>,
+    pub(crate) lattice_id: Arc<str>,
+    pub(crate) host_id: Arc<str>,
 }
 
 impl wasmcloud_provider_sdk::Provider for Provider {
@@ -172,6 +166,7 @@ impl wasmcloud_provider_sdk::Provider for Provider {
         });
         self.links
             .lock()
+            .instrument(trace_span!("insert_link"))
             .await
             .entry(target_id)
             .or_default()
@@ -188,73 +183,6 @@ impl wasmcloud_provider_sdk::Provider for Provider {
             .await
             .get_mut(target_id)
             .map(|links| links.remove(link_name));
-        Ok(())
-    }
-}
-
-impl crate::wasmbus::Host {
-    #[instrument(level = "debug", skip_all)]
-    pub(crate) async fn start_http_server_provider(
-        &self,
-        tasks: &mut JoinSet<()>,
-        link_definitions: impl IntoIterator<Item = InterfaceLinkDefinition>,
-        provider_xkey: XKey,
-        host_config: HashMap<String, String>,
-        provider_id: &str,
-        host_id: &str,
-    ) -> anyhow::Result<()> {
-        match host_config.get("routing_mode").map(String::as_str) {
-            // Run provider in address mode by default
-            Some("address") | None => {}
-            // Run provider in path mode
-            Some("path") => bail!("path mode not supported by builtin yet"),
-            Some(other) => bail!("unknown routing_mode: {other}"),
-        };
-        let default_address = host_config
-            .get("default_address")
-            .map(|s| SocketAddr::from_str(s))
-            .transpose()
-            .context("failed to parse default_address")?
-            .unwrap_or_else(default_listen_address);
-
-        let (quit_tx, quit_rx) = broadcast::channel(1);
-        let commands = ProviderCommandReceivers::new(
-            Arc::clone(&self.rpc_nats),
-            &quit_tx,
-            &self.host_config.lattice,
-            provider_id,
-            provider_id,
-            host_id,
-        )
-        .await?;
-        let conn = ProviderConnection::new(
-            Arc::clone(&self.rpc_nats),
-            Arc::from(provider_id),
-            Arc::clone(&self.host_config.lattice),
-            host_id.to_string(),
-            host_config,
-            provider_xkey,
-            Arc::clone(&self.secrets_xkey),
-        )
-        .context("failed to establish provider connection")?;
-        let provider = Provider {
-            address: default_address,
-            components: Arc::clone(&self.components),
-            links: Mutex::default(),
-            host_id: Arc::from(host_id),
-            lattice_id: Arc::clone(&self.host_config.lattice),
-        };
-        for ld in link_definitions {
-            if let Err(e) = receive_link_for_provider(&provider, &conn, ld).await {
-                error!(
-                    error = %e,
-                    "failed to initialize link during provider startup",
-                );
-            }
-        }
-        tasks.spawn(async move {
-            handle_provider_commands(provider, &conn, quit_rx, quit_tx, commands).await
-        });
         Ok(())
     }
 }
