+++ crates/host/src/wasmbus/providers/http_server/path.rs
@@ -1,0 +1,3 @@
+pub(crate) struct Provider {}
+
+impl wasmcloud_provider_sdk::Provider for Provider {}
